Refactor component imports and add new scenes
